Refactor code structure to not over-expose subsystem specific stuff
